total mess, but cleaned up almost-zero-copy bit reader
