+++ pco/src/constants.rs
@@ -18,7 +18,8 @@ pub const BITS_TO_ENCODE_N_BINS: Bitlen = 15;
 // performance tuning parameters
 const WORDS_PER_OFFSET_UPPER_BOUND: usize = 5;
 pub const DECOMPRESS_BYTE_PADDING: usize = BYTES_PER_WORD * WORDS_PER_OFFSET_UPPER_BOUND;
-// pub const UNSIGNED_BATCH_SIZE: usize = 512;
+pub const MINIMAL_PADDING_BYTES: usize = 1 << 6;
+pub const DEFAULT_PADDING_BYTES: usize = 1 << 13;
 
 // native architecture info
 pub const WORD_SIZE: usize = usize::BITS as usize;
@@ -31,11 +32,13 @@ pub const MAX_COMPRESSION_LEVEL: usize = 12;
 pub const MAX_DELTA_ENCODING_ORDER: usize = 7;
 pub const MAX_ENTRIES: usize = 1 << 24;
 pub const MAX_AUTO_DELTA_COMPRESSION_LEVEL: usize = 6;
+pub const MAX_SUPPORTED_PRECISION: Bitlen = 128;
 
 // defaults
 pub const DEFAULT_COMPRESSION_LEVEL: usize = 8;
 
 // other
+pub const FULL_BIN_BATCH_SIZE: usize = 128;
 pub const FULL_BATCH_SIZE: usize = 256;
 pub const ANS_INTERLEAVING: usize = 4;
 
@@ -57,5 +60,17 @@ mod tests {
       BITS_TO_ENCODE_DELTA_ENCODING_ORDER,
       MAX_DELTA_ENCODING_ORDER,
     );
+  }
+
+  #[test]
+  fn test_bin_batch_fits_in_default_padding() {
+    let max_bytes_per_bin = 2 * MAX_SUPPORTED_PRECISION as usize / 8 + 4;
+    assert!(FULL_BIN_BATCH_SIZE * max_bytes_per_bin < DEFAULT_PADDING_BYTES);
+  }
+
+  #[test]
+  fn test_batch_fits_in_default_padding() {
+    let max_bytes_per_num = MAX_SUPPORTED_PRECISION as usize / 8 + 2;
+    assert!(FULL_BATCH_SIZE * max_bytes_per_num < DEFAULT_PADDING_BYTES);
   }
 }
