+++ pco/src/latent_batch_decompressor.rs
@@ -5,12 +5,10 @@ use crate::ans::AnsState;
 use crate::bin::BinDecompressionInfo;
 use crate::bit_reader::BitReader;
 use crate::page_metadata::PageLatentMetadata;
-use crate::constants::{
-  Bitlen, ANS_INTERLEAVING, BYTES_PER_WORD, FULL_BATCH_SIZE, WORD_BITLEN, WORD_SIZE,
-};
+use crate::constants::{Bitlen, ANS_INTERLEAVING, BYTES_PER_WORD, FULL_BATCH_SIZE, WORD_BITLEN, WORD_SIZE, DEFAULT_PADDING_BYTES};
 use crate::data_types::UnsignedLike;
 use crate::errors::PcoResult;
-use crate::{ans, bits, ChunkLatentMetadata};
+use crate::{ans, bit_reader, bits, ChunkLatentMetadata};
 
 const MAX_ANS_SYMBOLS_PER_WORD: usize = 4;
 
@@ -54,8 +52,6 @@ impl<U: UnsignedLike> State<U> {
 #[derive(Clone, Debug)]
 pub struct LatentBatchDecompressor<U: UnsignedLike> {
   // known information about the latent latent in this chunk
-  max_bits_per_ans: Bitlen,
-  max_bits_per_offset: Bitlen,
   extra_words_per_offset: usize,
   infos: Vec<BinDecompressionInfo<U>>,
   maybe_constant_value: Option<U>,
@@ -73,13 +69,13 @@ impl<U: UnsignedLike> LatentBatchDecompressor<U> {
     needs_gcd: bool,
     is_trivial: bool,
   ) -> PcoResult<Self> {
-    let max_bits_per_ans = chunk_latent_meta.ans_size_log
-      - chunk_latent_meta
-        .bins
-        .iter()
-        .map(|bin| bin.weight.ilog2() as Bitlen)
-        .min()
-        .unwrap_or_default();
+    // let max_bits_per_ans = chunk_latent_meta.ans_size_log
+    //   - chunk_latent_meta
+    //   .bins
+    //   .iter()
+    //   .map(|bin| bin.weight.ilog2() as Bitlen)
+    //   .min()
+    //   .unwrap_or_default();
     let max_bits_per_offset = chunk_latent_meta
       .bins
       .iter()
@@ -100,8 +96,6 @@ impl<U: UnsignedLike> LatentBatchDecompressor<U> {
     let decoder = ans::Decoder::from_latent_meta(chunk_latent_meta)?;
 
     Ok(Self {
-      max_bits_per_ans,
-      max_bits_per_offset,
       extra_words_per_offset,
       infos,
       maybe_constant_value,
@@ -119,15 +113,16 @@ impl<U: UnsignedLike> LatentBatchDecompressor<U> {
 
   #[allow(clippy::needless_range_loop)]
   #[inline(never)]
-  fn unchecked_decompress_ans_tokens(&mut self, reader: &mut BitReader) {
-    let mut byte_idx = reader.loaded_byte_idx;
-    let mut bit_idx = reader.bits_past_ptr;
+  fn decompress_ans_tokens(&mut self, reader: &mut BitReader) {
+    let stream = reader.current_stream;
+    let mut stale_byte_idx = reader.stale_byte_idx;
+    let mut bit_idx = reader.bits_past_byte;
     let mut offset_bit_idx = 0;
     let mut state_idxs = self.state.state_idxs;
     for base_i in (0..FULL_BATCH_SIZE).step_by(MAX_ANS_SYMBOLS_PER_WORD) {
-      byte_idx += bit_idx as usize / 8;
+      stale_byte_idx += bit_idx as usize / 8;
       bit_idx %= 8;
-      let word = reader.unchecked_word_at(byte_idx);
+      let word = bit_reader::unchecked_word_at(stream, stale_byte_idx);
       // TODO this doesn't work on 32 bits or if MAX_ANS_SYMBOLS_PER_WORD != ANS_INTERLEAVING
       for j in 0..MAX_ANS_SYMBOLS_PER_WORD {
         let i = base_i + j;
@@ -140,66 +135,32 @@ impl<U: UnsignedLike> LatentBatchDecompressor<U> {
         state_idxs[j] = node.next_state_idx_base + state_offset;
       }
     }
-    reader.loaded_byte_idx = byte_idx;
-    reader.bits_past_ptr = bit_idx;
+    reader.stale_byte_idx = stale_byte_idx;
+    reader.bits_past_byte = bit_idx;
     self.state.state_idxs = state_idxs;
-  }
-
-  #[inline(never)]
-  fn decompress_ans(&mut self, reader: &mut BitReader, batch_size: usize) -> PcoResult<()> {
-    let mut state_idxs = self.state.state_idxs;
-    let mut offset_bit_idx = 0;
-    for i in 0..batch_size {
-      let j = i % ANS_INTERLEAVING;
-      let node = self.decoder.get_node(state_idxs[j]);
-      let state_offset = reader.read_small(node.bits_to_read)?;
-      let info = &self.infos[node.token as usize];
-      self.state.set_scratch(i, offset_bit_idx, info);
-      offset_bit_idx += info.offset_bits as usize;
-      state_idxs[j] = node.next_state_idx_base + state_offset;
-    }
-    self.state.state_idxs = state_idxs;
-    Ok(())
   }
 
   #[allow(clippy::needless_range_loop)]
   #[inline(never)]
-  fn unchecked_decompress_offsets<const MAX_EXTRA_WORDS: usize>(
+  fn decompress_offsets<const MAX_EXTRA_WORDS: usize>(
     &mut self,
     reader: &mut BitReader,
     dst: &mut [U],
   ) {
-    let base_bit_idx = reader.bit_idx();
+    let base_bit_idx = reader.current_stream_bit_idx();
+    let stream = reader.current_stream;
     for i in 0..FULL_BATCH_SIZE {
       let offset_bits = self.state.offset_bits_scratch[i];
       let bit_idx = base_bit_idx + self.state.offset_bits_csum_scratch[i];
+      let byte_idx = bit_idx / 8;
       let bits_past_byte = bit_idx as Bitlen % 8;
-      let mut byte_idx = bit_idx / 8;
-      let mut res = U::from_word(reader.unchecked_word_at(byte_idx) >> bits_past_byte);
-      let mut processed = min(offset_bits, WORD_BITLEN - 8 - bits_past_byte);
-      byte_idx += BYTES_PER_WORD - 1;
-
-      for _ in 0..MAX_EXTRA_WORDS {
-        res |= U::from_word(reader.unchecked_word_at(byte_idx)) << processed;
-        processed = min(offset_bits, processed + WORD_BITLEN);
-        byte_idx += BYTES_PER_WORD;
-      }
-
-      dst[i] = bits::lowest_bits(res, offset_bits);
+      dst[i] = bit_reader::unchecked_read_uint(stream, byte_idx, bits_past_byte, offset_bits);
     }
-    reader.seek_to(
-      base_bit_idx
+    let final_bit_idx = base_bit_idx
         + self.state.offset_bits_csum_scratch[FULL_BATCH_SIZE - 1]
-        + self.state.offset_bits_scratch[FULL_BATCH_SIZE - 1] as usize,
-    )
-  }
-
-  #[inline(never)]
-  fn decompress_offsets(&mut self, reader: &mut BitReader, dst: &mut [U]) -> PcoResult<()> {
-    for (i, x_dst) in dst.iter_mut().enumerate() {
-      *x_dst = reader.read_uint::<U>(self.state.offset_bits_scratch[i])?;
-    }
-    Ok(())
+        + self.state.offset_bits_scratch[FULL_BATCH_SIZE - 1] as usize;
+    reader.stale_byte_idx = final_bit_idx / 8;
+    reader.bits_past_byte = final_bit_idx as Bitlen % 8;
   }
 
   #[inline(never)]
@@ -236,29 +197,17 @@ impl<U: UnsignedLike> LatentBatchDecompressor<U> {
 
     let batch_size = dst.len();
     assert!(batch_size <= FULL_BATCH_SIZE);
+    reader.ensure_padded(DEFAULT_PADDING_BYTES);
 
-    // as long as there's enough compressed data available, we don't need checked operations
-    let is_full_batch = batch_size == FULL_BATCH_SIZE;
-    if is_full_batch && self.max_bits_per_ans as usize * FULL_BATCH_SIZE <= reader.bits_remaining()
-    {
-      self.unchecked_decompress_ans_tokens(reader);
-    } else {
-      self.decompress_ans(reader, batch_size)?;
-    }
+    self.decompress_ans_tokens(reader);
 
-    if is_full_batch
-      && self.max_bits_per_offset as usize * FULL_BATCH_SIZE <= reader.bits_remaining()
-    {
-      // this assertion saves some unnecessary specializations in the compiled assembly
-      assert!(self.extra_words_per_offset <= (U::PHYSICAL_BITS + 8) / WORD_SIZE);
-      match self.extra_words_per_offset {
-        0 => self.unchecked_decompress_offsets::<0>(reader, dst),
-        1 => self.unchecked_decompress_offsets::<1>(reader, dst),
-        2 => self.unchecked_decompress_offsets::<2>(reader, dst),
-        _ => panic!("invalid extra words per offset; a bug in pcodec"),
-      }
-    } else {
-      self.decompress_offsets(reader, dst)?;
+    // this assertion saves some unnecessary specializations in the compiled assembly
+    assert!(self.extra_words_per_offset <= (U::PHYSICAL_BITS + 8) / WORD_SIZE);
+    match self.extra_words_per_offset {
+      0 => self.decompress_offsets::<0>(reader, dst),
+      1 => self.decompress_offsets::<1>(reader, dst),
+      2 => self.decompress_offsets::<2>(reader, dst),
+      _ => panic!("invalid extra words per offset; a bug in pcodec"),
     }
 
     if self.needs_gcd {
