+++ pco/src/modes/classic.rs
@@ -24,20 +24,4 @@ impl<U: UnsignedLike> ConstMode<U> for ClassicMode {
   ) {
     bin.offset_bits = bits::bits_to_encode_offset(bin.upper - bin.lower);
   }
-
-  #[inline]
-  fn calc_offset(u: U, bin: &BinCompressionInfo<U>) -> U {
-    u - bin.lower
-  }
-
-  #[inline]
-  fn unchecked_decompress_unsigned(bin: &BinDecompressionInfo<U>, reader: &mut BitReader) -> U {
-    bin.lower + reader.unchecked_read_uint::<U>(bin.offset_bits)
-  }
-
-  #[inline]
-  fn decompress_unsigned(bin: &BinDecompressionInfo<U>, reader: &mut BitReader) -> PcoResult<U> {
-    let offset = reader.read_uint::<U>(bin.offset_bits)?;
-    Ok(bin.lower + offset)
-  }
 }
