+++ pco/src/modes/gcd.rs
@@ -61,21 +61,6 @@ impl<U: UnsignedLike> ConstMode<U> for GcdMode {
     bin.gcd = gcd;
     bin.offset_bits = bits::bits_to_encode_offset(max_offset);
   }
-
-  #[inline]
-  fn calc_offset(u: U, bin: &BinCompressionInfo<U>) -> U {
-    (u - bin.lower) / bin.gcd
-  }
-
-  #[inline]
-  fn unchecked_decompress_unsigned(bin: &BinDecompressionInfo<U>, reader: &mut BitReader) -> U {
-    bin.lower + reader.unchecked_read_uint::<U>(bin.offset_bits) * bin.gcd
-  }
-
-  #[inline]
-  fn decompress_unsigned(bin: &BinDecompressionInfo<U>, reader: &mut BitReader) -> PcoResult<U> {
-    Ok(bin.lower + reader.read_uint::<U>(bin.offset_bits)? * bin.gcd)
-  }
 }
 
 // fast if b is small, requires b > 0
